extract runner from singleton
